working project

+++ Program.cs
@@ -1,13 +1,20 @@
 ﻿using WypozyczalniaFilmow;
 
-Console.WriteLine("Wybierz dostep: pracownik, klient");
-string access = Console.ReadLine();
-if (access != "pracownik" && access != "klient")
-    Console.WriteLine("Podales zla wartosc");
+string access = "";
+while (access != "pracownik" && access != "klient")
+{
+    Console.WriteLine("Wybierz dostep: pracownik, klient");
+    access = Console.ReadLine();
+    if (access != "pracownik" && access != "klient")
+        Console.WriteLine("Podales zla wartosc");
+}
+
 bool end = true;
 while (end)
 {
+    Console.WriteLine("--------------------------------------------------------------------------------------------------");
     Console.WriteLine("1. Informacje o filmie \n" + "2. Dane klientow \n" + "3. Wypozyczenie \n" + "4. Stan magazynu \n" + "5. Wyciagnij z magazynu \n" + "6. Dodaj do magazynu");
+    Console.WriteLine("--------------------------------------------------------------------------------------------------");
     string selectedFunction = Console.ReadLine();
     if (selectedFunction == "end")
         break;
@@ -18,7 +25,9 @@ while (end)
             Movies movies = new Movies();
             DigitalMovies digitalMovies = new DigitalMovies();
             PhysicalMovies physicalMovies = new PhysicalMovies();
+            Console.WriteLine("--------------------------------------------------------------------------------------------------");
             Console.WriteLine("1. Wyswietl wszystkie filmy \n" + "2. Wyszukaj film \n" + "3. Wyszukaj wszystkie filmy cyfrowe \n" + "4. Wyszukaj wszystkie filmy na plytce");
+            Console.WriteLine("--------------------------------------------------------------------------------------------------");
             string selectMovies = Console.ReadLine();
             if (selectMovies == "end")
                 break;
@@ -51,7 +60,9 @@ while (end)
             {
                 Customers customers = new Customers();
                 PremiumCustomers premiumCustomers = new PremiumCustomers();
+                Console.WriteLine("--------------------------------------------------------------------------------------------------");
                 Console.WriteLine("1. Wyswietl wszystkich klientow \n" + "2. Wyszukaj klienta \n" + "3. Klienci premium");
+                Console.WriteLine("--------------------------------------------------------------------------------------------------");
                 string selectCustomers = Console.ReadLine();
                 if (selectCustomers == "end")
                     break;
@@ -84,7 +95,9 @@ while (end)
 
         case "3":
             Rental rent = new Rental();
+            Console.WriteLine("--------------------------------------------------------------------------------------------------");
             Console.WriteLine("1. Dodaj wypozyczenie \n" + "2. Wypisz wszystie wypozyczenia klienta \n" + "3. Usun wypozyczenie");
+            Console.WriteLine("--------------------------------------------------------------------------------------------------");
             string selectRent = Console.ReadLine();
             if (selectRent == "end")
                 break;
@@ -113,26 +126,36 @@ while (end)
                     break;
 
                 case "2":
-                    Console.WriteLine("Podaj imie i nazwisko klienta: ");
-                    string rentalClient = Console.ReadLine();
-                    if (rentalClient == "end")
-                        break;
-                    if (rentalClient != null)
-                        rent.findRentals(rentalClient);
+                    if (access == "pracownik")
+                    {
+                        Console.WriteLine("Podaj imie i nazwisko klienta: ");
+                        string rentalClient = Console.ReadLine();
+                        if (rentalClient == "end")
+                            break;
+                        if (rentalClient != null)
+                            rent.findRentals(rentalClient);
+                        else
+                            Console.WriteLine("Nie podano klienta");
+                    }
                     else
-                        Console.WriteLine("Nie podano klienta");
+                        Console.WriteLine("Nieautoryzowany dostep");
                     break;
 
                 case "3":
-                    Console.WriteLine("Podaj imie i nazwisko klienta: ");
-                    string clientNameRemove = Console.ReadLine();
-                    if (clientNameRemove == "end")
-                        break;
-                    Console.WriteLine("Podaj tytul filmu: ");
-                    string movieNameRemove = Console.ReadLine();
-                    if (movieNameRemove == "end")
-                        break;
-                    rent.deleteRental(movieNameRemove, clientNameRemove);
+                    if (access == "pracownik")
+                    {
+                        Console.WriteLine("Podaj imie i nazwisko klienta: ");
+                        string clientNameRemove = Console.ReadLine();
+                        if (clientNameRemove == "end")
+                            break;
+                        Console.WriteLine("Podaj tytul filmu: ");
+                        string movieNameRemove = Console.ReadLine();
+                        if (movieNameRemove == "end")
+                            break;
+                        rent.deleteRental(movieNameRemove, clientNameRemove);
+                    }
+                    else
+                        Console.WriteLine("Nieautoryzowany dostep");
                     break;
             }
 
@@ -144,21 +167,31 @@ while (end)
             break;
 
         case "5":
-            MovieStorage storedMoviesRemoving = new MovieStorage();
-            Console.WriteLine("Podaj tytul: ");
-            string titleRemove = Console.ReadLine();
-            if (titleRemove == "end")
-                break;
-            storedMoviesRemoving.removeMovie(titleRemove);
+            if (access == "pracownik")
+            {
+                MovieStorage storedMoviesRemoving = new MovieStorage();
+                Console.WriteLine("Podaj tytul: ");
+                string titleRemove = Console.ReadLine();
+                if (titleRemove == "end")
+                    break;
+                storedMoviesRemoving.removeMovie(titleRemove);
+            }
+            else
+                Console.WriteLine("Nieautoryzowany dostep");
             break;
 
         case "6":
-            MovieStorage storedMoviesAdding = new MovieStorage();
-            Console.WriteLine("Podaj tytul: ");
-            string titleAdd = Console.ReadLine();
-            if (titleAdd == "end")
-                break;
-            storedMoviesAdding.addMovie(titleAdd);
+            if (access == "pracownik")
+            {
+                MovieStorage storedMoviesAdding = new MovieStorage();
+                Console.WriteLine("Podaj tytul: ");
+                string titleAdd = Console.ReadLine();
+                if (titleAdd == "end")
+                    break;
+                storedMoviesAdding.addMovie(titleAdd);
+            }
+            else
+                Console.WriteLine("Nieautoryzowany dostep");
             break;
     }
 }
